Combine TaskSteam classes
